2 JPG images with text explanations
These are the initial Multiplayer space Ideas I have, as written in text and imagery in the JPGs uploaded. I will expand on this as I have ideas for it. This is simply a game I want to play, but willl probably not develop enought skill to create in my lifetime.